+++ CHANGES.rst
@@ -2,8 +2,8 @@ threedi-qgis-plugin changelog
 =============================
 
 
-0.8.3 (unreleased)
-------------------
+0.9 (2016-11-28)
+----------------
 
 - Update cumulative aggregation methods.